+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 All notable changes to this project will be documented in this file.
 
 --------
+## v1.1.0 (2019-10-17)
+### Added
+ - Functions for publish Data Sync Agent related events:
+  - `pubSyncPhysicalActivity(activity: any): Promise<void>`
+  - `pubSyncSleep(sleep: any): Promise<void>`
+  - `pubSyncWeight(weight: any): Promise<void>`
+  - `pubSyncLog(log: any): Promise<void>`
+  
+ - Functions for subscribe Data Sync Agent related events:
+  - `subFitbitAuthError(callback: (message: any) => void): Promise<void>`
+  - `subSyncPhysicalActivity(callback: (message: any) => void): Promise<void>`
+  - `subSyncSleep(callback: (message: any) => void): Promise<void>`
+  - `subSyncWeight(callback: (message: any) => void): Promise<void>`
+  - `subSyncLog(callback: (message: any) => void): Promise<void>`
+  
+ - Updated RPC Server and Client Roles
+  - Changed `getLogs(query: string): Promise<any>` to `getLogs(childId: string, query?: string): Promise<any>`;
+  - Now `query` parameter on all server and client functions is optional.
+
+### Fixed
+- Fixed RPC client listener that was not invalidating connection objects when it is closed.
+
 ## v1.0.5 (2019-09-30)
 ### Fixed
 - Fixed bug that did not remove events when close method was called;